requirements.txt
Faker==37.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
tzdata==2025.2